requirements.txt
ctransformers
sentence-transformers
pinecone-client
pinecone
langchain
flask
pypdf
python-dotenv
-e .